requirements.txt
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
PyJWT==2.8.0
psycopg==3.2.9
gunicorn==22.0.0
celery==5.4.0
redis==5.0.5
django-celery-beat==2.6.0
django-cors-headers==4.7.0
python-dotenv==1.1.1
whitenoise==6.9.0
requests